Update to v3.0.1-rc4: test-cloudinit.tf
The `cloudinit_cdrom_storage = "nvme"` option is obsolete in the latest Telmate Proxmox module, as it has been updated to the latest RC4 specifications. The script is now functioning again, but the latest version from master still needs to be *compiled* and manual installed as shown in the video!!
Currently, working with version v3.0.1-rc4.