Pinned k3s Version
Following the YouTube series i encountered a fatal problem with the new release of k3s being 1.27.x. 

The script for installing k3s is working with a pinned version (v1.26.10+k3s2). As a good measure i would either put them both on the stable channel or leave both of them at the same fixed version. 

Thanks for your work Jim and the great explanation!